Adding default SMCP yaml which uses Kiali and improving robustness
This commit is not chaning default behavior of OSSM installation script
but it allows to install Kiali for tests which require it. It also
improves robustness of OSSM install/delete scripts to make it work
reliably also on slower envs.

This change is required for a test John O'Hara is working on.